+++ src/core/response.cc
@@ -26,8 +26,9 @@ using namespace residue;
 
 const Response::StandardResponse Response::STANDARD_RESPONSES[] = {
     // do not use Session::PACKET_DELIMITER - sometimes resolves to nothing. Weird
-    { Response::StatusCode::STATUS_OK, "{r:0}\r\n\r\n" },
+    { Response::StatusCode::OK, "{r:0}\r\n\r\n" },
     { Response::StatusCode::CONTINUE, "{r:0}\r\n\r\n" },
     { Response::StatusCode::BAD_REQUEST, "{r:1}\r\n\r\n" },
+    { Response::StatusCode::INVALID_CLIENT, "{r:2}\r\n\r\n" },
 };