+++ src/tokenization/token-response.cc
@@ -39,7 +39,7 @@ TokenResponse::TokenResponse(const std::string& token,
                              int life,
                              const std::string& loggerId) :
     Response(),
-    m_status(StatusCode::STATUS_OK),
+    m_status(StatusCode::OK),
     m_token(token),
     m_loggerId(loggerId),
     m_life(life)
@@ -49,7 +49,7 @@ TokenResponse::TokenResponse(const std::string& token,
 
 TokenResponse::TokenResponse(bool) :
     Response(),
-    m_status(StatusCode::STATUS_OK),
+    m_status(StatusCode::OK),
     m_life(-1)
 {